refactor: User 관련 Service들을 application 아래로 이동
