+++ src/main/java/com/finnect/user/application/AuthenticateService.java
@@ -1,4 +1,4 @@
-package com.finnect.user.application.service;
+package com.finnect.user.application;
 
 import com.finnect.user.application.port.in.AuthenticateUseCase;
 import com.finnect.user.application.port.in.UserDetailsQuery;
